requirements.txt
fpdf
opencv-python
pdf2image
Pillow